vscode_snowflake_starter/requirements.txt
snowflake-connector-python>=3.11.0
python-dotenv>=1.0.1
